ai-knowledge-base-assistant/requirements.txt
langchain
langchain-google-genai
langchain-community
pydantic
python-dotenv
faiss-cpu